Worked on GUI, still so much to do
