+++ GUI/main.py
@@ -1,50 +1,19 @@
-# Author: Arthur De Neyer
-# Date: 01/03/2023
-# GitHub: GaecKo/MDPSaver
-
 import sys
-import platform
-from PySide6 import QtCore, QtGui, QtWidgets
-from PySide6.QtCore import (QCoreApplication, QPropertyAnimation, QDate, QDateTime, QMetaObject, QObject, QPoint, QRect, QSize, QTime, QUrl, Qt, QEvent)
-from PySide6.QtGui import (QBrush, QColor, QConicalGradient, QCursor, QFont, QFontDatabase, QIcon, QKeySequence, QLinearGradient, QPalette, QPainter, QPixmap, QRadialGradient)
-from PySide6.QtWidgets import *
+from PySide6.QtCore import Qt
+from PySide6.QtWidgets import QApplication
+from login import LoginPage
+from pages import MainWindow
 
-from ui_Main_App import Ui_MainWindow # ui file
-
-from ui_functions import * # functions file
-
-class MainWindow(QMainWindow):
-    def __init__(self):
-        QMainWindow.__init__(self)
-        self.ui = Ui_MainWindow()
-        self.ui.setupUi(self)
-
-        # toggle btn connection
-        self.ui.btn_toggle.clicked.connect(lambda: UIFunctions.toggleMenu(self, 250, True))
-
-        ## PAGES
-        ########################################################################
-
-        # Page 1
-        
-
-        ########################################################################
-        self.show()
-    
-    def connections(self):
-        # Pages connection: 
-        self.ui.btn_page_1.clicked.connect(lambda: self.ui.pages_widget.setCurrentWidget(self.ui.page_1))
-
-        # PAGE 2
-        self.ui.btn_page_2.clicked.connect(lambda: self.ui.pages_widget.setCurrentWidget(self.ui.page_2))
-
-        # PAGE 3
-        self.ui.btn_page_3.clicked.connect(lambda: self.ui.pages_widget.setCurrentWidget(self.ui.page_3))
-
-    def on_start(self):
-        pass
 
 if __name__ == "__main__":
     app = QApplication(sys.argv)
-    window = MainWindow()
+    app.setStyleSheet(open("styles.css").read())
+
+    # login_page = LoginPage()
+    main_window = MainWindow()
+
+    # login_page.successful_login.connect(main_window.show)
+    # login_page.show()
+    main_window.show()
+
     sys.exit(app.exec())
