+++ GUI/styles.css
@@ -1,18 +1,34 @@
 * {
     font-family: "Segoe UI";
+    color: #ecf0f1;
+    background-color: rgb(52, 53, 65);
 }
 
 QPushButton {
-    border: 1px solid #34495e;
-    background-color: #2c3e50;
+    background-color: rgb(64, 65, 79);
     color: #ecf0f1;
+    border: 0px;
     font-size: 14px;
-    padding: 5px 15px;
     border-radius: 5px;
 }
 
 QPushButton:hover {
     background-color: #34495e;
+}
+
+QGroupBox {
+    background-color: rgb(64, 65, 79);
+    border: none;
+}
+
+QVBoxLayout#leftPanel {
+    background-color: rgb(52, 53, 65);
+}
+
+QStackedWidget {
+    background-color: rgb(64, 65, 79);
+    border-radius: 20px;
+    border: none;
 }
 
 QPushButton:pressed {
@@ -23,10 +39,6 @@ QLineEdit {
     padding: 5px;
     border: 1px solid #bdc3c7;
     border-radius: 5px;
-}
-
-QMainWindow {
-    background-color: #ecf0f1;
 }
 
 QGroupBox {
